+++ src/main/java/com/wenxin2/marioverse/MarioverseCreativeTabs.java
@@ -160,6 +160,7 @@ public class MarioverseCreativeTabs {
             add(event, ItemRegistry.PIRANHA_PLANT_SPAWN_EGG);
             add(event, ItemRegistry.BOO_SPAWN_EGG);
             add(event, ItemRegistry.CHEEP_CHEEP_SPAWN_EGG);
+            add(event, ItemRegistry.SPINY_CHEEP_CHEEP_SPAWN_EGG);
         }
 
         if (event.getTabKey() == MARIOVERSE_BUILDING_BLOCKS_TAB.getKey() && !ConfigRegistry.DISABLE_MARIOVERSE_TABS.get()) {
@@ -718,6 +719,7 @@ public class MarioverseCreativeTabs {
                 addAfter(event, ItemRegistry.SNOW_POKEY_SPAWN_EGG, ItemRegistry.PIRANHA_PLANT_SPAWN_EGG);
                 addAfter(event, ItemRegistry.PIRANHA_PLANT_SPAWN_EGG, ItemRegistry.BOO_SPAWN_EGG);
                 addAfter(event, ItemRegistry.BOO_SPAWN_EGG, ItemRegistry.CHEEP_CHEEP_SPAWN_EGG);
+                addAfter(event, ItemRegistry.CHEEP_CHEEP_SPAWN_EGG, ItemRegistry.SPINY_CHEEP_CHEEP_SPAWN_EGG);
             }
 
             if (event.getTabKey() == CreativeModeTabs.BUILDING_BLOCKS) {