+++ changelog.md
@@ -5,6 +5,7 @@
   - Warm Cheep Cheeps spawn anywhere hot
   - Added Bucket of Cheep Cheep
   - Added Raw & Cooked Cheep Cheep
+- Added the Spiny Cheep Cheep
 - Added ability for Koopa Troopas to open doors
 - Added special named flags to the Creative Tabs